+++ tests/models/yolov5/test_yolov5.py
@@ -75,7 +75,7 @@ class ThisTester(ModelTester):
             os.remove(downloaded_file)
 
         subprocess.check_call([sys.executable, "-m", "pip", "uninstall", "-y", "opencv-python-headless"])
-        return model
+        return model.to(torch.bfloat16)
 
     def _load_inputs(self):
         # Image preprocessing
@@ -84,7 +84,7 @@ class ThisTester(ModelTester):
         transform = transforms.Compose([transforms.Resize((512, 512)), transforms.ToTensor()])
         img_tensor = [transform(image).unsqueeze(0)]
         batch_tensor = torch.cat(img_tensor, dim=0)
-        return batch_tensor
+        return batch_tensor.to(torch.bfloat16)
 
 
 @pytest.mark.parametrize(
